Pull tryOpenExternalFile out to a separate file so it can be more easily tested
